AQUAMIND/Aquamind/requirements.txt
# Aquamind 智能水处理系统依赖
# Python >= 3.10

# LangChain 核心
langchain>=1.2.0
langchain-core>=1.2.0
langchain-openai>=1.1.0
langchain-community>=0.4.0

# LangGraph (多智能体编排)
langgraph>=1.0.0
langgraph-prebuilt>=1.0.0

# 核心依赖
pydantic>=2.10.0
pydantic-settings>=2.10.0

# 数据处理
pandas>=2.0.0
numpy>=2.0.0

# HTTP 客户端
httpx>=0.28.0
aiohttp>=3.9.0

# 实时数据采集
paho-mqtt>=2.0.0          # MQTT 客户端
psycopg2-binary>=2.9.0    # PostgreSQL 连接

# 日志和配置
python-dotenv>=1.0.0

# 可选：向量数据库 (知识库增强)
# chromadb>=0.4.0
# faiss-cpu>=1.7.0

# 开发依赖 (可选)
# pytest>=8.0.0
# pytest-asyncio>=0.23.0
# black>=24.0.0
# ruff>=0.4.0
